+++ src/json-rpc/src/Exception/Handler/HttpExceptionHandler.php
@@ -12,41 +12,6 @@ declare(strict_types=1);
 
 namespace Hyperf\JsonRpc\Exception\Handler;
 
-use Hyperf\Contract\StdoutLoggerInterface;
-use Hyperf\ExceptionHandler\ExceptionHandler;
-use Hyperf\ExceptionHandler\Formatter\FormatterInterface;
-use Psr\Http\Message\ResponseInterface;
-use Throwable;
-
-class HttpExceptionHandler extends ExceptionHandler
+class HttpExceptionHandler extends TcpExceptionHandler
 {
-    /**
-     * @var StdoutLoggerInterface
-     */
-    protected $logger;
-
-    /**
-     * @var FormatterInterface
-     */
-    protected $formatter;
-
-    public function __construct(StdoutLoggerInterface $logger, FormatterInterface $formatter)
-    {
-        $this->logger = $logger;
-        $this->formatter = $formatter;
-    }
-
-    public function handle(Throwable $throwable, ResponseInterface $response)
-    {
-        $this->logger->warning($this->formatter->format($throwable));
-
-        $this->stopPropagation();
-
-        return $response;
-    }
-
-    public function isValid(Throwable $throwable): bool
-    {
-        return true;
-    }
 }
