+++ src/json-rpc/src/Exception/Handler/TcpExceptionHandler.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+
+namespace Hyperf\JsonRpc\Exception\Handler;
+
+use Hyperf\Contract\StdoutLoggerInterface;
+use Hyperf\ExceptionHandler\ExceptionHandler;
+use Hyperf\ExceptionHandler\Formatter\FormatterInterface;
+use Psr\Http\Message\ResponseInterface;
+use Throwable;
+
+class TcpExceptionHandler extends ExceptionHandler
+{
+    /**
+     * @var StdoutLoggerInterface
+     */
+    protected $logger;
+
+    /**
+     * @var FormatterInterface
+     */
+    protected $formatter;
+
+    public function __construct(StdoutLoggerInterface $logger, FormatterInterface $formatter)
+    {
+        $this->logger = $logger;
+        $this->formatter = $formatter;
+    }
+
+    public function handle(Throwable $throwable, ResponseInterface $response)
+    {
+        $this->logger->warning($this->formatter->format($throwable));
+
+        $this->stopPropagation();
+
+        return $response;
+    }
+
+    public function isValid(Throwable $throwable): bool
+    {
+        return true;
+    }
+}
